workers/keyword_intelligence/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keyword-intelligence-agent"
version = "0.2.0"
description = "AI SEO Tool - Keyword Intelligence Agent"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "AI SEO Tool Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "pinecone-client>=3.0.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
